day 7

+++ src/net/deflox/dez7/Dez7Part2.kt
@@ -10,7 +10,7 @@ fun main() {
         lines.add(it.toCharArray().toMutableList())
     }
 
-    val count = mutableListOf<Int>()
+    val count = mutableListOf<Long>()
     lines[2].forEach { char ->
         if (char == '^') count.add(1)
         else count.add(0)
@@ -19,8 +19,8 @@ fun main() {
     for (i in 2..lines.size - 2 step 2) {
         lines[i].forEachIndexed { charIndex, char ->
             if (char == '^') {
-                count[charIndex - 1] = if (count[charIndex - 1] == 0) count[charIndex] else count[charIndex - 1] + count[charIndex]
-                count[charIndex + 1] = if (count[charIndex + 1] == 0) count[charIndex] else count[charIndex + 1] + count[charIndex]
+                count[charIndex - 1] = if (count[charIndex - 1] == 0L) count[charIndex] else count[charIndex - 1] + count[charIndex]
+                count[charIndex + 1] = if (count[charIndex + 1] == 0L) count[charIndex] else count[charIndex + 1] + count[charIndex]
                 count[charIndex] = 0
             }
         }
